# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Jimu 后端框架 API - 提供用户认证、权限管理、角色管理、系统监控等功能",
        "title": "Jimu API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tenant-plans": {
            "get": {
                "description": "分页返回全部套餐定义。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户运营"
                ],
                "summary": "获取套餐列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认 20，最大 100）",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功，返回分页套餐列表",
                        "schema": {
                            "$ref": "#/definitions/contract.PageResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "description": "创建租户套餐（资源上限，0 表示不限）。编码全局唯一，统一转小写存储且不可修改。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户运营"
                ],
                "summary": "创建套餐",
                "parameters": [
                    {
                        "description": "套餐信息（编码、名称与各项上限）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.CreatePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功，返回套餐信息",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "套餐编码已存在",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/tenant-plans/{id}": {
            "put": {
                "description": "更新套餐名称与各项上限；编码不可修改。修改后对使用该套餐的租户立即生效。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户运营"
                ],
                "summary": "更新套餐",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套餐 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的字段（为空表示不修改）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.UpdatePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "套餐不存在",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "description": "删除未被任何租户使用的套餐；仍被租户使用时返回 409，避免租户静默失去配额约束。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户运营"
                ],
                "summary": "删除套餐",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套餐 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "404": {
                        "description": "套餐不存在",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "套餐仍被租户使用",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/tenants": {
            "get": {
                "description": "分页获取租户列表。支持按 ID、编码、创建时间排序。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "获取租户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认 20，最大 100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段（支持 id、code、created_at，默认 id）",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方向（asc 或 desc，默认 desc）",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功，返回分页租户列表",
                        "schema": {
                            "$ref": "#/definitions/contract.PageResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误（如无效的排序字段）",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "description": "创建新租户。编码全局唯一，仅允许字母、数字、短横线和下划线（统一转小写存储）；创建后编码不可修改。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "创建租户",
                "parameters": [
                    {
                        "description": "租户信息（编码和名称）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.CreateTenantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功，返回租户信息",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "参数错误（编码格式无效）",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "租户编码已存在",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/tenants/usage": {
            "get": {
                "description": "返回当前租户的套餐与各资源用量（用户/角色/API Key）及上限，用于容量自查；未分配套餐时 plan 为 null 且不限量。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户运营"
                ],
                "summary": "获取当前租户用量",
                "responses": {
                    "200": {
                        "description": "成功，返回套餐与用量",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/tenants/{id}": {
            "get": {
                "description": "根据租户 ID 获取租户详细信息。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "获取租户详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "租户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功，返回租户信息",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "参数错误（ID 格式无效）",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "租户不存在",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "description": "更新指定租户的名称或状态。编码不可修改。默认租户可改名但不可删除。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "更新租户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "租户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的租户信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.UpdateTenantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "租户不存在",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "description": "软删除指定租户。默认租户受保护，不可删除。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "删除租户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "租户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "参数错误（ID 格式无效）",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "租户不存在",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "默认租户不可删除",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/tenants/{id}/plan": {
            "put": {
                "description": "为指定租户分配套餐；plan_id=0 表示取消套餐（取消后不再受配额限制）。超限时不影响既有数据，只阻止继续创建。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户运营"
                ],
                "summary": "为租户分配套餐",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "租户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "套餐 ID（0=取消套餐）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.AssignPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分配成功",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "套餐不存在",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "application.AssignPlanRequest": {
            "type": "object",
            "properties": {
                "plan_id": {
                    "type": "integer"
                }
            }
        },
...
        "application.CreatePlanRequest": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1
                },
                "max_api_keys": {
                    "description": "0=不限",
                    "type": "integer",
                    "minimum": 0
                },
                "max_roles": {
                    "description": "0=不限",
                    "type": "integer",
                    "minimum": 0
                },
                "max_users": {
                    "description": "0=不限",
                    "type": "integer",
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                }
            }
        },
...
        "application.CreateTenantRequest": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 64
                },
                "name": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1
                }
            }
        },
...
        "application.UpdatePlanRequest": {
            "type": "object",
            "properties": {
                "max_api_keys": {
                    "type": "integer",
                    "minimum": 0
                },
                "max_roles": {
                    "type": "integer",
                    "minimum": 0
                },
                "max_users": {
                    "type": "integer",
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                }
            }
        },
...
        "application.UpdateTenantRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            }
        },
...
        "contract.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "contract.PageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "details": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}